requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
Flask==3.1.3
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.18.1
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openvino==2026.0.0
openvino-telemetry==2025.2.0
packaging==26.0
paho-mqtt==2.1.0
pandas==3.0.2
pillow==12.1.1
plotly==6.6.0
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==7.34.1
psutil==7.2.2
pyarrow==23.0.1
pydeck==0.9.1
pymodbus==3.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
streamlit-aggrid==1.2.1.post2
sympy==1.14.0
tenacity==9.1.4
toml==0.10.2
torch==2.11.0+cpu
torchvision==0.26.0+cpu
tornado==6.5.5
typing_extensions==4.15.0
ultralytics==8.4.33
ultralytics-thop==2.0.18
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.7
